backend/requirement.txt
alembic==1.0.11
aniso8601==7.0.0
asn1crypto==0.24.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
numpy==1.17.0
pandas==0.25.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.4
